Major repo clean-up and object refactor
