--- CaliPytion/modified/__init__.py
@@ -1,17 +1,0 @@
-from .calibrator import Calibrator
-from .standard import Standard
-from .sample import Sample
-from .model import Model
-from .parameter import Parameter
-from .spectrum import Spectrum
-
-__doc__ = ""
-
-__all__ = [
-    "Calibrator",
-    "Standard",
-    "Sample",
-    "Model",
-    "Parameter",
-    "Spectrum",
-]
