--- CaliPytion/schemes/CaliPytion_metadata.json
@@ -1,141 +1,0 @@
-{
-  "docstring": "Data model for (meta-) data of standard curves and spectra for spectrophotometric data.",
-  "enums": [
-    "TemperatureUnits",
-    "TemperatureUnits",
-    "ConcentrationUnits"
-  ],
-  "external": {},
-  "Calibration": {
-    "attributes": {
-      "reactant_id": {
-        "description": "Unique identifier of the calibrated reactant."
-      },
-      "date": {
-        "description": "Date when the calibration data was measured"
-      },
-      "pH": {
-        "description": "pH of solution."
-      },
-      "temperature": {
-        "description": "Temperature during calibration."
-      },
-      "temperature_unit": {
-        "description": "Temperature unit."
-      },
-      "device": {
-        "description": "Device object, containing information about the analytic device."
-      },
-      "standard": {
-        "description": "Standard data of a substance.",
-        "multiple": "True"
-      },
-      "spectrum": {
-        "description": "Spectrum data of a substance."
-      },
-      "result": {
-        "description": "Contains calculated concentrations and information on the fitted calibration equation to calculate the concentrations."
-      }
-    },
-    "docstring": null
-  },
-  "Device": {
-    "attributes": {
-      "manufacturer": {
-        "description": "Device manufacturer."
-      },
-      "model": {
-        "description": "Model name of the device."
-      },
-      "software_version": {
-        "description": "Software version of the device."
-      }
-    },
-    "docstring": null
-  },
-  "Standard": {
-    "attributes": {
-      "wavelength": {
-        "description": "Detection wavelength."
-      },
-      "concentration": {
-        "multiple": "True",
-        "description": "Concentration of the reactant."
-      },
-      "concentration_unit": {
-        "description": "Concentration unit."
-      },
-      "absorption": {
-        "multiple": "True",
-        "description": "Measured absorption, corresponding to the applied concentration of the reactant."
-      }
-    },
-    "docstring": "Description of a standard curve.\n"
-  },
-  "Spectrum": {
-    "attributes": {
-      "concentration": {
-        "description": "Concentration of the reactant at which the spectrum was recorded."
-      },
-      "wavelength": {
-        "multiple": "True",
-        "description": "Wavelengths used for detection."
-      },
-      "concentration_unit": {
-        "description": "Concentration unit."
-      },
-      "absorption": {
-        "multiple": "True",
-        "description": "Measured absorption, corresponding to detection wavelengths."
-      }
-    },
-    "docstring": null
-  },
-  "Result": {
-    "attributes": {
-      "concentration": {
-        "description": "Calculated concentration, based on the calibration model.",
-        "multiple": "True"
-      },
-      "calibration_model": {
-        "description": "Utilized calibration model and its parameters, leading to the calculated concentrations."
-      }
-    },
-    "docstring": null
-  },
-  "Model": {
-    "attributes": {
-      "name": {
-        "description": "Name of the calibration model."
-      },
-      "equation": {
-        "description": "Equation of the calibration model."
-      },
-      "parameters": {
-        "description": "Parameters of the calibration model equation.",
-        "multiple": "True"
-      }
-    },
-    "docstring": null
-  },
-  "Parameter": {
-    "attributes": {
-      "name": {
-        "description": "Name of the parameter."
-      },
-      "value": {
-        "description": "Value of the parameter."
-      }
-    },
-    "docstring": null
-  },
-  "Series": {
-    "attributes": {
-      "values": {
-        "multiple": "True",
-        "description": "Series representing an array of value"
-      }
-    },
-    "docstring": null
-  }
-}
