--- CaliPytion/schemes/Calibration data-model_metadata.json
@@ -1,100 +1,0 @@
-{
-  "docstring": "Data model handling reactant-standard data for UV-Vis photo-spectrometry.",
-  "enums": [
-    "TemperatureUnits",
-    "TemperatureUnits",
-    "ConcentrationUnits"
-  ],
-  "external": {},
-  "Calibration": {
-    "attributes": {
-      "reactant_id": {
-        "description": "Unique identifier of the calibrated reactant."
-      },
-      "date": {
-        "description": "Date when the calibration data was measured"
-      },
-      "pH": {
-        "description": "pH of solution."
-      },
-      "temperature": {
-        "description": "Temperature during calibration."
-      },
-      "temperature_unit": {
-        "description": "Temperature unit."
-      },
-      "device": {
-        "description": "Device object, containing information about the analytic device."
-      },
-      "standard": {
-        "description": "Standard data of a substance.",
-        "multiple": "True"
-      },
-      "spectrum": {
-        "description": "Spectrum data of a substance."
-      }
-    },
-    "docstring": null
-  },
-  "Device": {
-    "attributes": {
-      "manufacturer": {
-        "description": "Device manufacturer."
-      },
-      "model": {
-        "description": "Model name of the device."
-      },
-      "software_version": {
-        "description": "Software version of the device."
-      }
-    },
-    "docstring": null
-  },
-  "Standard": {
-    "attributes": {
-      "wavelength": {
-        "description": "Detection wavelength."
-      },
-      "concentration": {
-        "multiple": "True",
-        "description": "Concentration of the reactant."
-      },
-      "concentration_unit": {
-        "description": "Concentration unit."
-      },
-      "absorption": {
-        "multiple": "True",
-        "description": "Measured absorption, corresponding to the applied concentration of the reactant."
-      }
-    },
-    "docstring": "Description of a standard curve.\n"
-  },
-  "Spectrum": {
-    "attributes": {
-      "concentration": {
-        "description": "Concentration of the reactant at which the spectrum was recorded."
-      },
-      "wavelength": {
-        "multiple": "True",
-        "description": "Wavelengths used for detection."
-      },
-      "concentration_unit": {
-        "description": "Concentration unit."
-      },
-      "absorption": {
-        "multiple": "True",
-        "description": "Measured absorption, corresponding to detection wavelengths."
-      }
-    },
-    "docstring": null
-  },
-  "Series": {
-    "attributes": {
-      "values": {
-        "multiple": "True",
-        "description": "Series representing an array of value"
-      }
-    },
-    "docstring": null
-  }
-}
