--- CaliPytion/schemes/calipytion_schema.md
@@ -1,78 +1,0 @@
-```mermaid
-classDiagram
-    Calibrator *-- Standard
-    Calibrator *-- CalibrationModel
-    Standard *-- Sample
-    Standard *-- CalibrationModel
-    CalibrationModel *-- CalibrationRange
-    CalibrationModel *-- FitStatistics
-    CalibrationModel *-- Parameter
-    
-    class Calibrator {
-        +Standard standard
-        +float[0..*] concentrations
-        +UnitClass conc_unit
-        +float[0..*] signals
-        +CalibrationModel[0..*] models
-        +float cutoff
-    }
-    
-    class Standard {
-        +string species_id
-        +string name
-        +float wavelength
-        +Sample[0..*] samples
-        +float ph
-        +float temperature
-        +UnitClass temperature_unit
-        +datetime created
-        +CalibrationModel model_result
-    }
-    
-    class Sample {
-        +float concentration
-        +UnitClass conc_unit
-        +float signal
-    }
-    
-    class CalibrationModel {
-        +string name
-        +string equation
-        +Parameter[0..*] parameters
-        +boolean was_fitted
-        +CalibrationRange calibration_range
-        +FitStatistics statistics
-    }
-    
-    class CalibrationRange {
-        +float conc_lower
-        +float conc_upper
-        +float signal_lower
-        +float signal_upper
-    }
-    
-    class FitStatistics {
-        +float aic
-        +float bic
-        +float r2
-        +float rmsd
-    }
-    
-    class Parameter {
-        +string name
-        +float value
-        +float init_value
-        +float standard_error
-        +float lower_bound
-        +float upper_bound
-    }
-    
-    class Spectrum {
-        +string species_id
-        +string name
-        +float concentration
-        +float[0..*] conc_unit
-        +float[0..*] signals
-    }
-    
-```
