+++ requirements.txt
@@ -1,6 +1,6 @@
 ipython
 lmfit
-matplotlib
+plotly
 numpy
 pandas
 pydantic==1.8.2
